Move the romaji tag section into new namespace.
